+++ modules/minecraft/v1_21_1/v1_21_1-neoforge/src/main/kotlin/ru/lazyhat/compukters/impl/CompuktersMod.kt
@@ -39,6 +39,7 @@ import ru.lazyhat.compukters.impl.config.CompuktersServerConfig
 import ru.lazyhat.compukters.impl.fs.NeoForgeWorldFileSystemStores
 import ru.lazyhat.compukters.impl.ide.IdeClientBootstrap
 import ru.lazyhat.compukters.impl.ide.target.IdeTargetNetwork
+import ru.lazyhat.compukters.impl.peripheral.PeripheralConfiguratorNetwork
 import ru.lazyhat.compukters.impl.registry.CompuktersRegistry
 import ru.lazyhat.compukters.impl.terminal.TerminalNetwork
 import ru.lazyhat.compukters.lang.runtime.vm.JniRuntimeBackend
@@ -54,6 +55,7 @@ class CompuktersMod(
         CompuktersRegistry.register(eventBus)
         eventBus.addListener(TerminalNetwork::register)
         eventBus.addListener(IdeTargetNetwork::register)
+        eventBus.addListener(PeripheralConfiguratorNetwork::register)
         if (FMLEnvironment.dist == Dist.CLIENT) IdeClientBootstrap.register(eventBus)
         NeoForge.EVENT_BUS.addListener(NeoForgeWorldFileSystemStores::onLevelSave)
         NeoForge.EVENT_BUS.addListener(VmBenchmarkCommands::register)